Jormungandr: split the default scenario in two
The "simple" for all api except journeys, it is not mean to be use but
to be inherited for implementing the desired logic for journeys without
having to redifine all basic methods
